add many cfg files for cervix dataset add visualization during inference add support for cervix data(dataset, data path, evaluation fix path bugs, now you should make a soft link under ./dataset, eg. ln -s /data/cervix ./dataset/cervix add 'cocowrapper' for non-coco-style dataset to do coco evaluation add cache for gt annotation(for each dataset), predict result, dt annotation(for each checkpoint and score threshold) add tensorboard support for multi-gpu, record for different validation result add weight for classification loss

+++ maskrcnn_benchmark/engine/inference.py
@@ -66,47 +66,55 @@ def inference(
         expected_results=(),
         expected_results_sigma_tol=4,
         output_folder=None,
+        draw=False
 ):
-    # convert to a torch.device for efficiency
-    device = torch.device(device)
-    num_devices = get_world_size()
-    logger = logging.getLogger("maskrcnn_benchmark.inference")
-    dataset = data_loader.dataset
-    logger.info("Start evaluation on {} dataset({} images).".format(dataset_name, len(dataset)))
-    total_timer = Timer()
-    inference_timer = Timer()
-    total_timer.tic()
-    predictions = compute_on_dataset(model, data_loader, device, inference_timer)
-    # wait for all processes to complete before measuring the time
-    synchronize()
-    total_time = total_timer.toc()
-    total_time_str = get_time_str(total_time)
-    logger.info(
-        "Total run time: {} ({} s / img per device, on {} devices)".format(
-            total_time_str, total_time * num_devices / len(dataset), num_devices
+    # load saved result
+    if os.path.exists(os.path.join(output_folder, "predictions.pth")):
+        if not is_main_process():
+            return
+        predictions = torch.load(os.path.join(output_folder, "predictions.pth"),  map_location=torch.device("cpu"))
+    else:
+        # convert to a torch.device for efficiency
+        device = torch.device(device)
+        num_devices = get_world_size()
+        logger = logging.getLogger("maskrcnn_benchmark.inference")
+        dataset = data_loader.dataset
+        logger.info("Start evaluation on {} dataset({} images).".format(dataset_name, len(dataset)))
+        total_timer = Timer()
+        inference_timer = Timer()
+        total_timer.tic()
+        predictions = compute_on_dataset(model, data_loader, device, inference_timer)
+        # wait for all processes to complete before measuring the time
+        synchronize()
+        total_time = total_timer.toc()
+        total_time_str = get_time_str(total_time)
+        logger.info(
+            "Total run time: {} ({} s / img per device, on {} devices)".format(
+                total_time_str, total_time * num_devices / len(dataset), num_devices
+            )
         )
-    )
-    total_infer_time = get_time_str(inference_timer.total_time)
-    logger.info(
-        "Model inference time: {} ({} s / img per device, on {} devices)".format(
-            total_infer_time,
-            inference_timer.total_time * num_devices / len(dataset),
-            num_devices,
+        total_infer_time = get_time_str(inference_timer.total_time)
+        logger.info(
+            "Model inference time: {} ({} s / img per device, on {} devices)".format(
+                total_infer_time,
+                inference_timer.total_time * num_devices / len(dataset),
+                num_devices,
+            )
         )
-    )
 
-    predictions = _accumulate_predictions_from_multiple_gpus(predictions)
-    if not is_main_process():
-        return
+        predictions = _accumulate_predictions_from_multiple_gpus(predictions)
+        if not is_main_process():
+            return
 
-    if output_folder:
-        torch.save(predictions, os.path.join(output_folder, "predictions.pth"))
+        if output_folder:
+            torch.save(predictions, os.path.join(output_folder, "predictions.pth"))
 
     extra_args = dict(
         box_only=box_only,
         iou_types=iou_types,
         expected_results=expected_results,
         expected_results_sigma_tol=expected_results_sigma_tol,
+        draw=draw
     )
 
     return evaluate(dataset=dataset,
